Move deprecated reference loading out into service
