+++ lib/graphwerk/presenters/package.rb
@@ -24,9 +24,7 @@ module Graphwerk
 
       sig { returns(T::Array[String]) }
       def deprecated_references
-        return [] if !deprecated_references_file.exist?
-
-        (YAML.load_file(deprecated_references_file) || {}).keys.map do |reference|
+        DeprecatedReferencesLoader.new(@package, @root_path).load.map do |reference|
           Name.new(reference).node_name
         end
       end
@@ -42,14 +40,6 @@ module Graphwerk
       end
 
       private
-
-      DEPRECATED_REFERENCES_FILENAME = 'deprecated_references.yml'
-
-      sig { returns(Pathname) }
-      def deprecated_references_file
-        @deprecated_references_file = T.let(@deprecated_references_file, T.nilable(Pathname))
-        @deprecated_references_file ||= @root_path.join(@package.name, DEPRECATED_REFERENCES_FILENAME)
-      end
 
       sig { returns(Name) }
       def package_name
